added checks to migrations

+++ db/migrate/20120911151845_create_reminder_configurations.rb
@@ -1,6 +1,6 @@
 class CreateReminderConfigurations < ActiveRecord::Migration
   def self.up
-    create_table :reminder_configurations do |t|
+    create_table :reminder_configurations, :force => true do |t|
       t.column :days, :integer
       t.column :issue_status_selector, :string
       t.column :project_selector, :string
@@ -10,10 +10,10 @@ class CreateReminderConfigurations < ActiveRecord::Migration
       t.column :send_to_custom_user, :boolean
 
       t.timestamps
-    end
+    end unless table_exists? :reminder_configurations
   end
 
   def self.down
-    drop_table :reminder_configurations
+    drop_table :reminder_configurations if table_exists? :reminder_configurations
   end
 end
