+++ db/migrate/20120918093621_create_reminder_trackers.rb
@@ -3,10 +3,10 @@ class CreateReminderTrackers < ActiveRecord::Migration
     create_table :reminder_trackers do |t|
       t.column :reminder_configuration_id, :integer
       t.column :tracker_id, :integer
-    end
+    end unless table_exists? :reminder_trackers
   end
 
   def self.down
-    drop_table :reminder_trackers
+    drop_table :reminder_trackers if table_exists? :reminder_trackers
   end
 end
